remove unnecessary type for user_names_filter
