+++ insta_user_names-filter/main/main.go
@@ -1,9 +1,12 @@
 package main
 
 import (
-	filter "github.com/codeuniversity/smag-mvp/insta_user_names-filter"
+	"encoding/json"
+
+	"github.com/codeuniversity/smag-mvp/kafka/changestream"
 	"github.com/codeuniversity/smag-mvp/service"
 	"github.com/codeuniversity/smag-mvp/utils"
+	"github.com/segmentio/kafka-go"
 )
 
 func main() {
@@ -12,10 +15,29 @@ func main() {
 	changesTopic := utils.GetStringFromEnvWithDefault("KAFKA_CHANGE_TOPIC", "postgres.public.users")
 	namesTopic := utils.GetStringFromEnvWithDefault("KAFKA_NAME_TOPIC", "user_names")
 
-	f := filter.New(kafkaAddress, groupID, changesTopic, namesTopic)
+	f := changestream.NewFilter(kafkaAddress, groupID, changesTopic, namesTopic, filterChange)
 
 	service.CloseOnSignal(f)
 	waitUntilClosed := f.Start()
 
 	waitUntilClosed()
+}
+
+type user struct {
+	ID       int    `json:"id"`
+	UserName string `json:"user_name"`
+}
+
+func filterChange(m *changestream.ChangeMessage) ([]kafka.Message, error) {
+	if m.Payload.Op != "c" {
+		return nil, nil
+	}
+
+	u := &user{}
+	err := json.Unmarshal(m.Payload.After, u)
+	if err != nil {
+		return nil, err
+	}
+
+	return []kafka.Message{{Value: []byte(u.UserName)}}, nil
 }
